xdp: get source ip address in execute
Currently, we try to build source ip address if not specified by user in
two places: in execute and in `tx_loop`.
This PR moves all the code to one place simplifying the code.
